Refactor the energy labels EPI to use the database column
Instead of doing the calculations of the imputed EPI in Python,
we use the previously calculated epi_imputed column from the database.
This is cleaner and faster.

To save this in a csv as we did previously, it's probably easiest to not
use Python but to do it inside PostgreSQL itself.