Add table_name to index_name for add_index
This is necessary, because index names have to be unique across the schema,
which includes the other tables. Therefore, adding indexes on columns
from different tables with the same column name won't work: due to the
'IF NOT EXISTS' clause, only one of them will be created.

This led to confusing results, where it seemed that Postgres didn't want
to create any new indexes, while not giving any errors.
The existence of an index with the same name but on a different table
was actually the culprit.