Add support for RegionalModules
Now that we also have Regions as a seperate class,
we need modules that work on Regions specifically.
This prevents us from overloading Dwelling Modules
with regional functionalities, and will speed up
processing since we only need to process a region
once instead of processing it for every dwelling.

In addition, all Modules now take optional **kwargs.
The main reason was to support passing the regional_modules
variables to the RegionsModule, but this might come in handy
later on for other modules.

The implementation of the EnergyLabelRegionalModule is a stub
to show the design of these modules. New regional processing
options can be added later on, for e.g. 'wijken' and 'buurten'.