Add support for 'Buurt'
We use the Dutch word 'Buurt' (neighbourhood) since
it is a statistical classifier (used by a.o. CBS), and
the translation can clash with the other classifier 'Wijk'.

A new class 'Region' is added as a semantic parent for Buurt and PC6,
but it does not have any functionality yet.