Lower pipeline memory usage by deleting dwelling references after use
After a dwelling has been processed and saved, there are still two
references to the dwelling left:
- in the PC6 (self.dwelling)
- in the Buurt (self.dwellings)
These regions stay 'alive' since they are themselves referenced
from the RegionalModule (in the future even these references
could be deleted, but the memory pressure of the Regions is
way lower than those of the dwellings).

Thus, if we want to free up the memory used for a dwelling,
we need to delete the references from the Regions.
We can safely do this after all Placeholders in a Region
have been replaced by the actual dwellings.
Note that actually, the Placeholders only function is to keep
the data that has been collected before and to transfer it
to dwellings that are added. After that, the dwelling in a region
has no purpose. So we could in the future even decide to skip
the replacement.

This process works because all Region-level attributes are calculated
during the initializatoin of the Region. If this changes, then
this memory-freeing approach should be changed.

Bonus: instead of fetching the whole BAG query results in Python
memory, we use a server-side cursor. This saves ~1.2GB in the
beginning.